+++ Source/Game/Scripting/Interfaces/IPlayer/IPlayerOnLevelChanged.cs
@@ -8,6 +8,6 @@ namespace Game.Scripting.Interfaces.IPlayer
     // Called when a player's level changes (after the level is applied);
     public interface IPlayerOnLevelChanged : IScriptObject
     {
-        void OnLevelChanged(Player player, byte oldLevel);
+        void OnLevelChanged(Player player, uint oldLevel);
     }
 }
